Tests failing due to tests having same name -- didn't know that was a pytest issue
